gave up and used tables.
configured acl with casl

+++ packages/backend/src/snippets/snippets.controller.ts
@@ -6,14 +6,26 @@ import {
   Logger,
   Param,
   Post,
+  Request,
   Put,
+  UseGuards,
+  UnauthorizedException,
 } from '@nestjs/common';
+import { Request as Req } from 'express';
+import { JwtAuthGuard } from '@auth/jwt-auth.guard';
 import { SnippetsService } from './snippets.service';
+import { CaslAbilityFactory } from '@acl/casl-ability.factory';
 import { SnippetPostBodyDTO, SnippetPutBodyDTO } from './snippet-body.dto';
+import { ValidatedTokenUser } from '@auth/auth.types';
+import { Action } from '@auth/constants';
+import { Snippet } from '@entities/snippet.entity';
 
 @Controller('snippets')
 export class SnippetsController {
-  constructor(private snippetService: SnippetsService) {}
+  constructor(
+    private snippetService: SnippetsService,
+    private caslAbilityFactory: CaslAbilityFactory,
+  ) {}
 
   logger: Logger = new Logger(SnippetsController.name);
 
@@ -27,23 +39,37 @@ export class SnippetsController {
     return this.snippetService.findById(id);
   }
 
+  @UseGuards(JwtAuthGuard)
   @Post()
-  async createSnippet(@Body() { title, language, code }: SnippetPostBodyDTO) {
+  async createSnippet(
+    @Body() { title, language, code }: SnippetPostBodyDTO,
+    @Request() request: Req,
+  ) {
+    const user: any = request.user;
+    this.logger.log(JSON.stringify(user));
     try {
-      return this.snippetService.create(code, language, title);
+      return this.snippetService.create(code, language, user['userId'], title);
     } catch (error) {
       this.logger.error(error?.message ?? '');
       throw error;
     }
   }
 
+  @UseGuards(JwtAuthGuard)
   @Put(':snippetId')
   async updateSnippet(
     @Param('snippetId') id: string,
     @Body() { title, language, code }: SnippetPutBodyDTO,
+    @Request() request: Req,
   ) {
     {
       try {
+        const ability = this.caslAbilityFactory.createForUser(
+          request.user as ValidatedTokenUser,
+        );
+        if (!ability.can(Action.UPDATE, Snippet)) {
+          throw new UnauthorizedException();
+        }
         return this.snippetService.update(id, code, language, title);
       } catch (error) {
         this.logger.error(error?.message ?? '');
@@ -52,10 +78,17 @@ export class SnippetsController {
     }
   }
 
+  @UseGuards(JwtAuthGuard)
   @Delete(':snippetId')
-  async deleteSnippet(@Param('snippetId') id: string) {
+  async deleteSnippet(@Param('snippetId') id: string, @Request() request: Req) {
     {
       try {
+        const ability = this.caslAbilityFactory.createForUser(
+          request.user as ValidatedTokenUser,
+        );
+        if (!ability.can(Action.UPDATE, Snippet)) {
+          throw new UnauthorizedException();
+        }
         return this.snippetService.delete(id);
       } catch (error) {
         this.logger.error(error?.message ?? '');
